requirements.txt
fastapi
uvicorn
transformers==4.33.2
torch
torchvision
Pillow
